LeetCode-Java/src/main/java/com/enchanted/leetcode/LeetCode654.java
package com.enchanted.leetcode;

/**
 * @ClassName: LeetCode654
 * @Description: 654. 最大二叉树
 * @PackageName:com.enchanted.leetcode
 * @Author Enchanted
 * @Date 2023/11/23 10:25
 * @Version 1.0
 */


public class LeetCode654 {
    public TreeNode constructMaximumBinaryTree(int[] nums) {
        return construct(nums, 0, nums.length - 1);
    }

    public TreeNode construct(int[] nums, int left, int right) {
        if (left > right) {
            return null;
        }
        int best = left;
        for (int i = left + 1; i <= right; ++i) {
            if (nums[i] > nums[best]) {
                best = i;
            }
        }
        TreeNode node = new TreeNode(nums[best]);
        node.left = construct(nums, left, best - 1);
        node.right = construct(nums, best + 1, right);
        return node;
    }
}
class TreeNode {
    int val;
    TreeNode left;
    TreeNode right;

    TreeNode() {
    }

    TreeNode(int val) {
        this.val = val;
    }

    TreeNode(int val, TreeNode left, TreeNode right) {
        this.val = val;
        this.left = left;
        this.right = right;
    }
}
